+++ SDK/AppCenterCrashes/Microsoft.AppCenter.Crashes.Windows.Shared/Crashes.cs
@@ -4,6 +4,7 @@
 using Microsoft.AppCenter.Channel;
 using Microsoft.AppCenter.Crashes.Ingestion.Models;
 using Microsoft.AppCenter.Crashes.Utils;
+using Microsoft.AppCenter.Crashes.Windows.Shared.Utils;
 using Microsoft.AppCenter.Ingestion.Models;
 using Microsoft.AppCenter.Ingestion.Models.Serialization;
 using Microsoft.AppCenter.Utils;
@@ -24,6 +25,13 @@ namespace Microsoft.AppCenter.Crashes
         private static Crashes _instanceField;
 
         private const int MaxAttachmentsPerCrash = 2;
+
+        /// <summary>
+        /// Indicates if the app received a low memory warning in the last session.
+        /// </summary>
+        private bool _hasReceivedMemoryWarning;
+
+        internal const string PrefKeyMemoryWarning = Constants.KeyPrefix + "MemoryWarning";
 
         internal const string PrefKeyAlwaysSend = Constants.KeyPrefix + "CrashesAlwaysSend";
 
@@ -133,8 +141,14 @@ namespace Microsoft.AppCenter.Crashes
         // Task to get the last session error report, if one is found.
         private TaskCompletionSource<ErrorReport> _lastSessionErrorReportTaskSource;
 
+        /// <summary>
+        /// Memory warning helper.
+        /// </summary>
+        internal IMemoryWarningHelper _memoryWarningHelper;
+
         internal Crashes()
         {
+            _memoryWarningHelper = new MemoryWarningHelper();
             _unprocessedManagedErrorLogs = new Dictionary<Guid, ManagedErrorLog>();
             _errorReportCache = new ConcurrentDictionary<Guid, ErrorReport>();
         }
@@ -168,9 +182,12 @@ namespace Microsoft.AppCenter.Crashes
                     ChannelGroup.SendingLog += ChannelSendingLog;
                     ChannelGroup.SentLog += ChannelSentLog;
                     ChannelGroup.FailedToSendLog += ChannelFailedToSendLog;
+                    _memoryWarningHelper.MemoryWarning += OnMemoryWarning;
                 }
                 else if (!enabled)
                 {
+                    _memoryWarningHelper.MemoryWarning -= OnMemoryWarning;
+                    ApplicationSettings.Remove(PrefKeyMemoryWarning);
                     ApplicationLifecycleHelper.Instance.UnhandledExceptionOccurred -= OnUnhandledExceptionOccurred;
                     if (ChannelGroup != null)
                     {
@@ -221,7 +238,7 @@ namespace Microsoft.AppCenter.Crashes
 
         private async Task<bool> InstanceHasReceivedMemoryWarningInLastSessionAsync()
         {
-            return await Task.FromResult(false);
+            return await Task.FromResult(_hasReceivedMemoryWarning);
         }
 
         private Task ProcessPendingErrorsAsync()
@@ -279,6 +296,13 @@ namespace Microsoft.AppCenter.Crashes
                         RemoveAllStoredErrorLogFiles(log.Id);
                     }
                 }
+                var memoryWarning = ApplicationSettings.GetValue(PrefKeyMemoryWarning, false);
+                if (memoryWarning)
+                {
+                    _hasReceivedMemoryWarning = true;
+                    AppCenterLog.Debug(LogTag, "The application received a low memory warning in the last session.");
+                }
+                ApplicationSettings.Remove(PrefKeyMemoryWarning);
                 _lastSessionErrorReportTaskSource.SetResult(lastSessionErrorReport);
                 await SendCrashReportsOrAwaitUserConfirmationAsync().ConfigureAwait(false);
             });
@@ -451,6 +475,12 @@ namespace Microsoft.AppCenter.Crashes
             {
                 SentErrorReport?.Invoke(sender, new SentErrorReportEventArgs { Report = report });
             }
+        }
+
+        private void OnMemoryWarning(object sender, EventArgs e)
+        {
+            ApplicationSettings.SetValue(PrefKeyMemoryWarning, true);
+            AppCenterLog.Debug(LogTag, "The application received a low memory warning.");
         }
 
         private void ChannelFailedToSendLog(object sender, FailedToSendLogEventArgs e)